+++ packages/edition-slices/src/tiles/tile-ai/styles/index.js
@@ -1,0 +1,14 @@
+import { spacing } from "@times-components/styleguide";
+
+const styles = {
+  container: {
+    alignItems: "center",
+    paddingBottom: spacing(3),
+    paddingTop: spacing(2)
+  },
+  imageContainer: {
+    width: "100%"
+  }
+};
+
+export default styles;