+++ packages/edition-slices/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.41.0](https://github.com/newsuk/times-components/compare/@times-components/edition-slices@0.40.1...@times-components/edition-slices@0.41.0) (2019-03-18)
+
+
+### Features
+
+* Tablet - Comment Lead and Cartoon ([#1806](https://github.com/newsuk/times-components/issues/1806)) ([f54dd32](https://github.com/newsuk/times-components/commit/f54dd32))
+
+
+
+
+
 ## [0.40.1](https://github.com/newsuk/times-components/compare/@times-components/edition-slices@0.40.0...@times-components/edition-slices@0.40.1) (2019-03-18)