+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.19.4](https://github.com/newsuk/times-components/compare/@times-components/pages@1.19.3...@times-components/pages@1.19.4) (2019-03-18)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 ## [1.19.3](https://github.com/newsuk/times-components/compare/@times-components/pages@1.19.2...@times-components/pages@1.19.3) (2019-03-18)